+++ src/main/java/com/cibertec/Proyecto_Bodega/serviceImpl/MarcaService.java
@@ -23,7 +23,7 @@ public class MarcaService {
         mrc.save(m);
     }
 
-    public void eliminarMarca(Integer codMarc){
+    public void eliminarMarca(Long codMarc){
         mrc.deleteById(codMarc);
     }
 }
